+++ docs/user_manual.md
@@ -907,20 +907,38 @@ and stores the result in JSON file.
 The task types supported are `classification`, `detection`, and `segmentation`.
 
 The validation result contains
-- annotation statistics based on the task type
-- validation reports, such as
+- `annotation statistics` based on the task type
+- `validation reports`, such as
     - items not having annotations
     - items having undefined annotations
     - imbalanced distribution in class/attributes
     - too small or large values
-- summary
+- `summary`
 
 Usage:
+- There are five configurable parameters for validation
+    - `few_samples_thr` : threshold for giving a warning for minimum number of samples per class
+    - `imbalance_ratio_thr` : threshold for giving imbalance data warning
+    - `far_from_mean_thr` : threshold for giving a warning that data is far from mean
+    - `dominance_ratio_thr` : threshold for giving a warning bounding box imbalance
+    - `topk_bins` : ratio of bins with the highest number of data to total bins in the histogram
 
 ``` bash
 datum validate --help
 
-datum validate -p <project dir> <task_type>
+datum validate -p <project dir> -t <task_type> -- \
+    -fs <few_samples_thr> \
+    -ir <imbalance_ratio_thr> \
+    -m <far_from_mean_thr> \
+    -dr <dominance_ratio_thr> \
+    -k <topk_bins>
+```
+
+Example : give warning when imbalance ratio of data with classification task over 40
+
+``` bash
+datum validate -p prj-cls -t classification -- \
+    -ir 40
 ```
 
 Here is the list of validation items(a.k.a. anomaly types).
